backend/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==4.0.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.11
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
